requirements.txt
Flask
google-genai
gunicorn
PyYAML
requests
